# app/api/zulu/top-rank/route.ts
import { NextRequest, NextResponse } from 'next/server'
import { getZuluTopByRank } from '@/lib/leaders/zulu'

export async function GET(request: NextRequest) {
  try {
    const { searchParams } = new URL(request.url)
    const limit = parseInt(searchParams.get('limit') || '50')
    
    const traders = await getZuluTopByRank(limit)
    return NextResponse.json(traders)
  } catch (error: any) {
    console.error('Error fetching top Zulu traders:', error)
    return NextResponse.json(
      { error: error.message || 'Failed to fetch top traders' },
      { status: 500 }
    )
  }
}