# lib/openapi/investment-agent-openapi.json
{
  "openapi": "3.1.0",
...
  "servers": [
    {
      "url": "https://timetravel.investments/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/zulu/search": {
      "get": {
        "tags": [
          "Zulu"
        ],
        "summary": "Search Zulu traders",
        "description": "Search for Zulu traders with performance filters",
        "parameters": [
          {
            "name": "minRoi",
            "in": "query",
            "description": "Minimum ROI percentage",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minWinRate",
            "in": "query",
            "description": "Minimum Win Rate percentage",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxDrawdown",
            "in": "query",
            "description": "Maximum Drawdown percentage",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isEa",
            "in": "query",
            "description": "Filter by EA (Expert Advisor) usage",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max results (default: 50)",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/zulu/top-rank": {
      "get": {
        "tags": [
          "Zulu"
        ],
        "summary": "Get top ranked traders",
        "description": "Get list of top ranked Zulu traders",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Max results (default 50)",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
...
  },
...
}